[#358] partition components interops: opt-ins in dedicated files, fixes, in-code documentation
