+++ libs/ag/examples/01_overview_demo.cpp
@@ -1,6 +1,7 @@
-#include <csl/typeinfo.hpp> // optional: gives csl::ag::io::typenamed clean type names (e.g. "int")
-#define CSL_AG__ENABLE_STD_FORMAT_SUPPORT 1
 #include <csl/ag.hpp>
+#include <csl/ag/formatting/format.hpp>   // opt-in: std::formatter support (+ csl::ag::io::to_string)
+#include <csl/typeinfo.hpp>               // bridge prerequisite (explicit, for godbolt raw-URL include order)
+#include <csl/ag/formatting/typeinfo.hpp> // opt-in: gives csl::ag::io::typenamed clean type names (e.g. "int")
 #include <iostream> // std::print might not be available yet: use `std::cout << std::format(...)`
 
 struct S { char c; int i; };
@@ -20,9 +21,9 @@ auto main() -> int {
     constexpr auto format_options = indexed | typenamed | indented;
     std::cout << std::format("{}", value | format_options); // equivalent to std::println("{:xit}", value)
 
-    // legacy alternative: CSL_AG__ENABLE_IOSTREAM_SUPPORT
+    // alternative: #include <csl/ag/formatting/ostream.hpp>
     // std::cout << "value: " << format_options << value << '\n';
 
-    // other alternative: CSL_AG__ENABLE_FMT_SUPPORT
+    // alternative: #include <csl/ag/formatting/fmt.hpp>
     // fmt::println("{:xit}", value)
 }
