+++ libs/ag/examples/23_formatting_std_format_simple.cpp
@@ -1,7 +1,7 @@
-#include <csl/typeinfo.hpp> // optional: gives csl::ag::io::typenamed clean type names (e.g. "int")
-#define CSL_AG__ENABLE_STD_FORMAT_SUPPORT 1
-
 #include <csl/ag.hpp>
+#include <csl/ag/formatting/format.hpp>   // opt-in: std::formatter support (+ csl::ag::io::to_string)
+#include <csl/typeinfo.hpp>               // bridge prerequisite (explicit, for godbolt raw-URL include order)
+#include <csl/ag/formatting/typeinfo.hpp> // opt-in: gives csl::ag::io::typenamed clean type names (e.g. "int")
 #include <format>
 #include <iostream>
