+++ libs/ag/tests/odr_mixed_plain.cpp
@@ -1,0 +1,25 @@
+// ODR sanity (fmt-style feature-header contract):
+// this TU does NOT include the typeinfo bridge - see odr_mixed_bridge.cpp for the rationale and the contract being verified.
+#include <csl/ag.hpp>
+#include <csl/ag/formatting/format.hpp>
+
+#include <format>
+#include <typeindex>
+
+#include <catch2/catch_test_macros.hpp>
+
+// NOLINTBEGIN(*-avoid-do-while)
+// NOLINTBEGIN(*-avoid-magic-numbers)
+
+namespace test::ag::odr {
+    struct plain { char c; };
+}
+
+TEST_CASE("odr-mixed: plain TU - typenamed uses the <typeindex> fallback", "[ag][formatting][odr]") {
+    using namespace csl::ag::io;
+    const auto expected = std::format("{{{}: 'A'}}", std::type_index(typeid(char)).name());
+    CHECK(to_string<typenamed>(test::ag::odr::plain{ .c = 'A' }) == expected);
+}
+
+// NOLINTEND(*-avoid-magic-numbers)
+// NOLINTEND(*-avoid-do-while)
